pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "browser-use>=0.7.9",
    "click>=8.3.0",
    "fastapi>=0.116.2",
    "jinja2>=3.1.6",
    "openai>=1.108.1",
    "playwright>=1.55.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "rich>=14.1.0",
    "router>=0.1",
    "stringzilla>=4.0.13",
    "uvicorn>=0.36.0",
]
